requirements.txt
# ML Agent for Demand Shift Detection and Non-Moving Inventory
# Python 3.11+

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0

# Time Series & Statistics
scipy>=1.10.0
statsmodels>=0.14.0

# Machine Learning
scikit-learn>=1.3.0

# API Framework
fastapi>=0.100.0
uvicorn>=0.23.0

# Data Validation
pydantic>=2.0.0

# Visualization
plotly>=5.15.0
matplotlib>=3.7.0

# Database
sqlite-utils>=3.35.0
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0

